+++ src/test/java/model/PortalAppsPage.java
@@ -1,6 +1,5 @@
 package model;
 
-import model.portal.PortalBaseIndexPage;
 import org.openqa.selenium.WebDriver;
 
 public class PortalAppsPage extends PortalBaseIndexPage {
